wave-5 banked + follow-ups: E12 both grids in (CondDiff EOG 0.453/0.888 loses to SimpleCNN 0.321/0.934 on the EEGDfus grid — honest), E3b/E4 banked with the uncomfortable finding that our 1-ch transplant underperforms identity on SSED (cc 0.49 vs 0.70, bridge exonerated at cc 1.000); F4 ICA arm built for all 9 subjects; F5 Table-I matrix script (their EEG_Net_8_Stack + frozen recipe from their own ckpt path string) raw+ica submitted; E5 PLV chained on the waveform-saving re-evals; DS-DDPM probe label-dtype fix (their loop casts label.float()); EEGDfus eval unwraps the DDPM-wrapper state dict

+++ scripts/paper_final/dsddpm_repro.py
@@ -94,7 +94,7 @@ def probe() -> None:
     configs.epochs = 1
     configs.init()
     batch = next(iter(configs.data_loader))
-    xb, yb = batch[0].to(device), batch[1].to(device)
+    xb, yb = batch[0].to(device), batch[1].float().to(device)
     checks["batch_shape"] = list(xb.shape)
     xb = torch.permute(xb, (0, 3, 2, 1))     # their train(): unet2d_overlap.py:528
     checks["model_input_shape"] = list(xb.shape)
